+++ src/controllers/cart.controller.ts
@@ -1,23 +1,13 @@
 // src/cart/cart.controller.ts
 
-import type { Prisma } from '@prisma/client';
 import type { NextFunction, Request, Response } from 'express';
 
-import { STATUS_CODE } from '../constants/constant.js';
+import { MESSAGE, STATUS_CODE } from '../constants/constant.js';
 import type { PatchCartInput } from '../dtos/cart.dto.js';
 import { getCartItemWithDetails, getCartWithItems } from '../repositories/cart.repository.js';
+import { CartResponse } from '../serializes/cart.serialize.js';
 import * as cartService from '../services/cart.service.js';
 import { HttpException } from '../utils/http-exception.js';
-
-type SizeDetail = { ko?: string; en?: string; [key: string]: unknown };
-
-const toSizeDetail = (value: Prisma.JsonValue | null): SizeDetail | null => {
-  if (!value || typeof value !== 'object' || Array.isArray(value)) return null;
-  return value as SizeDetail;
-};
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-type CartItemWithDetails = any;
 
 class CartController {
   private getUserIdOr401(req: Request): string {
@@ -26,78 +16,6 @@ class CartController {
       throw HttpException.unauthorized('인증이 필요합니다.');
     }
     return user.id;
-  }
-
-  private transformCartItem(item: CartItemWithDetails) {
-    return {
-      id: item.id,
-      cartId: item.cartId,
-      productId: item.productId,
-      sizeId: item.sizeId,
-      quantity: item.quantity,
-      createdAt: item.createdAt,
-      updatedAt: item.updatedAt,
-      product: item.product
-        ? {
-            id: item.product.id,
-            storeId: item.product.storeId,
-            name: item.product.name,
-            price: item.product.price,
-            image: item.product.image,
-            createdAt: item.product.createdAt,
-            updatedAt: item.product.updatedAt,
-            reviewsRating: item.product.reviewsRating,
-            categoryId: item.product.categoryId,
-            content: item.product.content,
-            isSoldOut: item.product.isSoldOut,
-            discountRate: item.product.productDiscounts?.[0]?.discountRate ?? 0,
-            discountStartTime: item.product.productDiscounts?.[0]?.discountStartTime ?? null,
-            discountEndTime: item.product.productDiscounts?.[0]?.discountEndTime ?? null,
-            store: item.product.store
-              ? {
-                  id: item.product.store.id,
-                  userId: item.product.store.userId,
-                  name: item.product.store.name,
-                  address: item.product.store.address,
-                  phoneNumber: item.product.store.phoneNumber,
-                  content: item.product.store.content,
-                  image: item.product.store.image,
-                  createdAt: item.product.store.createdAt,
-                  updatedAt: item.product.store.updatedAt,
-                  detailAddress: item.product.store.detailAddress,
-                }
-              : null,
-            stocks: item.product.stocks?.map(
-              (s: {
-                id: string;
-                productId: string;
-                sizeId: string;
-                quantity: number;
-                size: { id: string; name: string; sizeDetail: Prisma.JsonValue } | null;
-              }) => ({
-                id: s.id,
-                productId: s.productId,
-                sizeId: s.sizeId,
-                quantity: s.quantity,
-                size: s.size
-                  ? {
-                      id: s.size.id,
-                      name: s.size.name,
-                      size: toSizeDetail(s.size.sizeDetail),
-                    }
-                  : null,
-              }),
-            ),
-          }
-        : null,
-      size: item.size
-        ? {
-            id: item.size.id,
-            name: item.size.name,
-            size: toSizeDetail(item.size.sizeDetail),
-          }
-        : null,
-    };
   }
 
   // POST /api/cart : 장바구니 생성 (또는 기존 카트 반환)
@@ -109,7 +27,7 @@ class CartController {
       if (!cart) {
         throw new HttpException({
           status: STATUS_CODE.INTERNAL_SERVER_ERROR,
-          message: '장바구니 생성에 실패했습니다.',
+          message: MESSAGE.cartCreationFailed,
         });
       }
 
@@ -145,12 +63,12 @@ class CartController {
 
       const cartItemId = req.params.cartItemId;
       if (!cartItemId) {
-        throw HttpException.badRequest('잘못된 요청입니다.');
+        throw HttpException.badRequest(MESSAGE.badRequest);
       }
 
       const deletedCount = await cartService.removeCartItem(userId, cartItemId);
       if (!deletedCount) {
-        throw HttpException.notFound('장바구니에 아이템이 없습니다.');
+        throw HttpException.notFound(MESSAGE.cartItemNotFound);
       }
 
       return res.status(STATUS_CODE.NO_CONTENT).send();
@@ -168,7 +86,7 @@ class CartController {
 
       // 장바구니가 없으면 404
       if (!cart) {
-        throw HttpException.notFound('장바구니를 찾을 수 없습니다.');
+        throw HttpException.notFound(MESSAGE.cartNotFound);
       }
 
       // 아이템이 없으면 빈 배열로 반환
@@ -179,7 +97,7 @@ class CartController {
         buyerId: cart.userId,
         createdAt: cart.createdAt,
         updatedAt: cart.updatedAt,
-        items: items.map((item) => this.transformCartItem(item)),
+        items: items.map((item) => CartResponse.transformCartItem(item)),
       });
     } catch (error) {
       return next(error);
@@ -193,48 +111,15 @@ class CartController {
 
       const { cartItemId } = req.params;
       if (!cartItemId) {
-        throw HttpException.badRequest('잘못된 요청입니다.');
+        throw HttpException.badRequest(MESSAGE.badRequest);
       }
 
       const item = await getCartItemWithDetails(userId, cartItemId);
       if (!item) {
-        throw HttpException.notFound('장바구니에 아이템이 없습니다.');
+        throw HttpException.notFound(MESSAGE.cartItemNotFound);
       }
 
-      // Swagger 스펙에 맞춘 간소화된 응답
-      const response = {
-        id: item.id,
-        cartId: item.cartId,
-        productId: item.productId,
-        sizeId: item.sizeId,
-        quantity: item.quantity,
-        createdAt: item.createdAt,
-        updatedAt: item.updatedAt,
-        product: item.product
-          ? {
-              id: item.product.id,
-              storeId: item.product.storeId,
-              name: item.product.name,
-              price: item.product.price,
-              image: item.product.image,
-              createdAt: item.product.createdAt,
-              updatedAt: item.product.updatedAt,
-              discountRate: item.product.productDiscounts?.[0]?.discountRate ?? 0,
-              discountStartTime: item.product.productDiscounts?.[0]?.discountStartTime ?? null,
-              discountEndTime: item.product.productDiscounts?.[0]?.discountEndTime ?? null,
-            }
-          : null,
-        cart: item.cart
-          ? {
-              id: item.cart.id,
-              buyerId: item.cart.userId,
-              createdAt: item.cart.createdAt,
-              updatedAt: item.cart.updatedAt,
-            }
-          : null,
-      };
-
-      return res.status(STATUS_CODE.OK).json(response);
+      return res.status(STATUS_CODE.OK).json(CartResponse.transformCartItemSimple(item));
     } catch (error) {
       return next(error);
     }
