+++ src/dtos/cart.dto.ts
@@ -2,7 +2,7 @@
 import { z } from 'zod';
 
 export const patchCartDto = z.object({
-  productId: z.string(),
+  productId: z.string().min(1, '상품 ID가 필요합니다'),
   sizes: z
     .array(
       z.object({
